計算式作成

+++ db/migrate/20200904124825_create_inventories.rb
@@ -1,9 +1,9 @@
 class CreateInventories < ActiveRecord::Migration[6.0]
   def change
     create_table :inventories do |t|
-      t.integer :order
-      t.integer :use
-      t.integer :stock
+      t.float :order
+      t.float :use
+      t.float :stock
       t.date :date, null: false
       t.references :item, null: false, foreign_key: true
       t.references :user, null: false, foreign_key: true
